Allow reconfigure NEO if enabled
Check the REVLibError not just for succes, but also for a kCannotPersistParametersWhileEnabled. The new configuration does get accepted, but wouldn't survive a brownout. If you wanted this to work, just transmit the NTvalue while robot is in disabled state.
Also, fixed the NTDoubles for steering gains, which had lambda's modifying the drive gains.