Motor Initialization Command Sequence Creation (#6)
* Create New Command

Each time the setup is run, a new command sequence needs to be created. The scheduler seems to delete the command sequence once it is done running, so it becomes undefined behavior to schedule the command again.

* Allow reconfigure NEO if enabled

Check the REVLibError not just for succes, but also for a kCannotPersistParametersWhileEnabled. The new configuration does get accepted, but wouldn't survive a brownout. If you wanted this to work, just transmit the NTvalue while robot is in disabled state.
Also, fixed the NTDoubles for steering gains, which had lambda's modifying the drive gains.

* Combined scheduleSetup()

Combined separate motor type setup commands into one scheduleSetup() function with a switch case.