Created software hard stop logic code.
Finished writing the code for the implementation of a software hard stop, with min and max values set. Values need to be set in the future once arm motor is calibrated. Fixed CANID logic so that we are referencing a Analog ID as well as a CANID. Import cleanups as well.