Intake and Outtake Command Sequences
Added the Command sequences that control the intake of both Algae and Coral (One Command Sequence). Also added the command sequence that controls the outtake (expel) of both algae and coral. Did this by checking if the beam is broken or if the velocity is lower than 10 or the output current is greater than 40. If the Beam is broken, we know we have a coral but if one or the other of the two conditions is met, we know we have an algae. This is because the NEO motor will start to slow down as it keeps sucking in the algae, which will slow the velocity or jump the output current which is why we check for both. Did some testing but didn't get far :/