Detect Motor Types on CAN Bus
Make testing on breadboards easier by detecting motor types if they exist and simulating them if they don't. Gives warning if motors are simulated when FMS is connected or in Driver Station Test mode. This also applies to the gyro. If it's not detected, it's put into simulation mode.

Add support for old Pigeon IMU v1.1.
For breadboard testing and replacement flexibility allow either Pigeon 2.0 or Pigeon 1.1