Auto Drive Cancel
Only cancel auto driving at the moment the input increases beyond a value with hysteresis. This should allow the auto driving to be queued while driving full speed.
Clears the pose queue in driveReefOffset, so if driver presses button twice, it won't queue it twice and cause a missed coral.