How did this get so big?
In autoconstants, there were two very similarly named variables for the reef center. We happened to pick the wrong one, so it was not driving to the correct center align distance for algae in auto. Renamed the variables to be clearly different.
Minor things: combine instantCommands to remove 20ms delays between them.
Don't allow elevator to move if intake is running, but do allow the buttonbox to move the target. Before it would ignore the dial command if intakeRunning, making it seem like it wasn't working.
Add WAYPOINT and add waitUntil for AUTOPOSE.
Fix Path Scorer to get out of obstacles better by allowing the two nearest nodes to not be penalized.
Add debounce to dial0, so that it doesn't go down and up when switching the dial.
Maybe eliminated the staticReefOffset! (well just moved it into the front camera offset)
Faster Algae Intake Speed